refactor: prepare structure for migrate didweb service
