+++ apps/vs-agent/src/controllers/admin/agent/index.ts
@@ -1,0 +1,1 @@
+export * from './VsAgentController'
