+++ apps/vs-agent/src/controllers/admin/credentials/index.ts
@@ -1,0 +1,3 @@
+export * from './CreateRevocationRegistryDto'
+export * from './CredentialTypeController'
+export * from './CredentialTypeDto'
