+++ apps/vs-agent/src/controllers/admin/index.ts
@@ -1,0 +1,7 @@
+export * from './agent'
+export * from './connections'
+export * from './credentials'
+export * from './invitation'
+export * from './message'
+export * from './presentations'
+export * from './vcauthn'
