+++ apps/vs-agent/src/controllers/admin/message/index.ts
@@ -1,0 +1,3 @@
+export * from './MessageController'
+export * from './MessageService'
+export * from './services'
