+++ apps/vs-agent/src/controllers/admin/message/services/index.ts
@@ -1,0 +1,3 @@
+export * from './CoreMessageService'
+export * from './MessageServiceFactory'
+export * from './RedisMessageService'
