+++ apps/vs-agent/src/controllers/admin/presentations/index.ts
@@ -1,0 +1,1 @@
+export * from './PresentationsController'
