+++ apps/vs-agent/src/controllers/admin/vcauthn/index.ts
@@ -1,0 +1,3 @@
+export * from './OobInvitationDto'
+export * from './PresentProofCreateDto'
+export * from './VCAuthNController'
