+++ apps/vs-agent/src/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from './admin'
+export * from './didcomm'
